+++ processor/transform/tables-to-jsx.ts
@@ -1,4 +1,4 @@
-import { Parents, Table, TableCell, Text } from 'mdast';
+import { Node, Paragraph, Parents, Table, TableCell, Text } from 'mdast';
 import { visit, EXIT } from 'unist-util-visit';
 import { Transform } from 'mdast-util-from-markdown';
 
@@ -17,11 +17,18 @@ const alignToStyle = (align: 'left' | 'center' | 'right' | null) => {
   };
 };
 
+const isTableCell = (node: Node) => ['tableHead', 'tableCell'].includes(node.type);
+
 const visitor = (table: Table, index: number, parent: Parents) => {
   let hasFlowContent = false;
 
   const tableCellVisitor = (cell: TableCell) => {
-    if (!phrasing(cell.children[0])) {
+    const content =
+      cell.children.length === 1 && cell.children[0].type === 'paragraph'
+        ? (cell.children[0] as unknown as Paragraph).children[0]
+        : cell.children[0];
+
+    if (!phrasing(content)) {
       hasFlowContent = true;
       return EXIT;
     }
@@ -34,8 +41,11 @@ const visitor = (table: Table, index: number, parent: Parents) => {
     });
   };
 
-  visit(table, 'tableCell', tableCellVisitor);
-  if (!hasFlowContent) return;
+  visit(table, isTableCell, tableCellVisitor);
+  if (!hasFlowContent) {
+    table.type = 'table';
+    return;
+  }
 
   const styles = table.align.map(alignToStyle);
 
@@ -98,8 +108,10 @@ const visitor = (table: Table, index: number, parent: Parents) => {
   parent.children[index] = jsx;
 };
 
+const isTable = (node: Node) => ['table', 'tableau'].includes(node.type);
+
 const tablesToJsx = (): Transform => tree => {
-  visit(tree, 'table', visitor);
+  visit(tree, isTable, visitor);
 
   return tree;
 };